mainloop-worker: rename thread_id to thread_index when used as a term
The "thread_id" value had dual uses:

  1) at one point it was 1 based, so that 0 meant uninitialized state
  2) in other cases it was 0 based, so we could use it as an array index

This patch makes the two uses distinct. The "id" term is only referred
to the 1 based value, 0 meaning uninitialized. In all other cases we
are referring to "thread_index" to better reflect that this is a 0
based value. In the latter case, -1 means uninitialized.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>