dbld/builddeps: workaround setuptools 66 PEP440 enforcement
pip can fail with setuptools>=66 if there are system packages
installed which do notcomply with PEP440:

https://github.com/pypa/setuptools/issues/3772#issuecomment-1398809718

As long as our supported distros have invalidly named packages,
this will be an issue.

Either debian package maintainers start to conform PEP440 versioning,
or setuptools loosens this strictness or this workaround sticks for a
while.

Either way, we need to see make or CI green, so until one of the above
happens I think we should go with this workaround.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>