diskq: store disk-buf-size in disk-queue header
This will be useful for a dqtool fix (see the next couple of commits)
and later on for increasing and decreasing the disk-queue file size.

Note that with this commit we only store the initially set disk-buf-size
and take care of old disk-queue files, which we created when we did not
store it yet. Currently we do not base any logic on it. That will come
in the next couple of commits.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>