diskq: use disk-buf-size stored in disk-queue file and reject changing it
There is a bug in dqtool which was caused by 2 things:
  1. We changed the condition of the cursor wrapping from file size to
     disk-buf-size().
  2. We do not store the disk-buf-size in the disk-queue header.

There is no way to tell where we should wrap without knowing the
disk-buf-size before loading the disk-queue.

The previous commit prepared a fix by storing the initial disk-buf-size
in the header. Now we use that value for the wrap condition instead
the user given one.

We can now detect whether the user tries to change the disk-buf-size
of an existing disk-queue file. For now we will reject that change and
go with the original disk-buf-size, which is a bugfix itself, as
before this commit we did not realize that with a config change and
might have tried to overread/overwrite the disk-queue file.

We can implement the disk-buf-size change in the future in a safe way,
however it is not the scope of this patch set.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>