qdisk: add disk-buf-size autodetect for non-wrapped v2 disk-queue files
In theory this should only be done if disk-buf-size() is not set
when loading the disk-queue file == dqtool cat or diskq-source().

If the cursors are not wrapped, we can work with any disk-buf-size(),
which is larger, than the write head. Additionally, we can assume, that
the disk-buf-size() was set to at least the size of the file, which is
only not true, if the last message written was started just before the
original disk-buf-size() and ended after it. This means that in a non
wrapped case, if we use the current file size as the disk-buf-size(),
we either use a smaller than originally set, or in lucky cases exactly
equal or similarly (un)lucky cases nearly one message larger.

I think this is an acceptable compromise, but either way, we are not
going to write more messages to the disk-queue file, so it is not a big
deal.

Unfortunately, we could not deal with wrapped disk-queue files even
before this patch set, so I did not fix it. We could write some
heuristics, so that the first short read indicates that we should wrap
instead, but I do not really want to complicate the wrap logic, if it
is not absolutely necessary.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>