metrics-probe: add skeleton
This kind of parser counts messages passing through, based on the
metadata of each message. It creates labeled stats counters based on the
fields of the message. Both the key and labels can be set in the config,
the values of the labels can be templated.

This commit creates the skeleton for this parser, the implementation
will be done in the following commits.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>