stats: add labeling capability
This commit adds labels to stats cluster keys.

These labels are similar to the labels in the Prometheus data model:
https://prometheus.io/docs/concepts/data_model/

The (id, labels) pair identifies a new-style key. The order of the labels
should not matter, and a label with an empty value should be equivalent to
a label non-existing, but the current implementation is rudimentary and
does not take these considerations into account.

Signed-off-by: László Várady <laszlo.varady@anno.io>