cfg-parser: fix potential NULL deref in the error display functions
If we are trying to print a location in the source file that does not
exist (e.g. if we use an incorrect line number for instance). Incorrect
error numbers are not really possible, unless we specify them using
@line (upcoming patch) or the file is truncated after being read.

I stumbled into this because of the first (e.g. @line).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>