cfg-lexer: avoid using borrowed "level" instance from CFG_LTYPE
The include_stack is not always changing in sync with the grammar rules,
unfortunately, so it might happen that an CFG_LTYPE instance is still
processed while its CfgIncludeLevel has already been destroyed.

In this case yylloc->level would point to zeroed out data. Not very
helpful when we are trying to use that location information later.

In the new model, we are storing a pointer to the filename, along with
row/column information. The filename is interned using g_intern_string,
so the pointer is not freed. This leaves us with some extra memory
allocated for each filename we pull in as we read those files.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>