metrics-probe: add labels() option
Why do we need to limit the number of labels?

We need allocated memory for the formatted labels, so we can create
a `StatsClusterKey` from them. We could store a buffer for them in the
class, but that would make it stateful, which means we need to lock it
in the `process()` function
Alternatively we could `malloc()` it in each `process()` call, but it
is costly.

If the number of labels is limited, we can allocate them on the stack
in the `process()` function, which means we save a per message
`malloc()` and manage to keep the parser stateless.

The downside is that we put a limit on it, but I don't think that having
more than 32 labels is sensible.

Everything mentioned in this commit message will become cleaner in the
next commits, where the labels get used.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>