dbld: do not install pip packages during image build
Except KIRA, which needs python2 and python3 at the same time and it is
not running tests with our prepared build or install venv.

Debian declines to install pip packages to system paths, because of
https://peps.python.org/pep-0668/
The solution to this is to install pip packages to a venv.

At this point I am not sure that we need pip packages installed to the
dbld images, as we have build and install time venvs. The only benefit is
that we ship these pip packages with the image, and those packages can be
used for anything not syslog-ng related even, but I think it is not enough
of a reason to keep them there.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>